Implemented taking damage
Former-commit-id: dc9c0fb03a96d787da5bb1a1ebd35baba98919b6

+++ Assets/Scripts/Core/IHealth.cs
@@ -1,5 +1,5 @@
 namespace BoM.Core {
 	public interface IHealth {
-		void TakeDamage(int damage, ISkill skill, IPlayer caster);
+		void TakeDamage(float damage, ISkill skill, IPlayer caster);
 	}
 }
